ref: move search functions to search.utils.ts
