+++ src/utils/data.utils.ts
@@ -1,6 +1,3 @@
-import { escapeRegExp } from "./regex.utils";
-import { isValidRegExp } from "../utils/regex.utils";
-
 export function isValidInput(data: unknown): boolean {
   const isObject = (val: unknown) => typeof val === "object" && val !== null;
   return (
@@ -121,113 +118,6 @@ export function getPaginated(options: {
   }
 
   return rows;
-}
-
-const meetsCriteria = (
-  inputText: string,
-  pattern: string,
-  options: {
-    matchCase: boolean;
-    matchWord: boolean;
-    useRegExp: boolean;
-  },
-): boolean => {
-  const { matchCase = false, matchWord = false, useRegExp = false } = options;
-
-  const flags = matchCase ? "g" : "gi";
-
-  if (matchWord) {
-    pattern = `\\b(${pattern})\\b`;
-  }
-
-  if (!useRegExp) {
-    inputText = escapeRegExp(inputText);
-  }
-
-  return new RegExp(pattern, flags).test(inputText);
-};
-
-export function search(options: {
-  rows: NormalisedRow[];
-  searchQuery: string;
-  useRegExp: boolean;
-  matchCase: boolean;
-  matchWord: boolean;
-}): NormalisedRow[] {
-  let { searchQuery = "" } = options;
-
-  if (!searchQuery) return options.rows;
-
-  const excludeFieldMatches: [string, string][] = [];
-  searchQuery = searchQuery.replace(/(?<!\w)-\w+:\w+/g, (match) => {
-    const [field, value] = match.replace("-", "").split(":");
-    excludeFieldMatches.push([field, value]);
-    return "";
-  });
-
-  const includeFieldMatches: [string, string][] = [];
-  searchQuery = searchQuery.replace(/\b\w+:\w+\b/g, (match) => {
-    const [field, value] = match.split(":");
-    includeFieldMatches.push([field, value]);
-    return "";
-  });
-
-  const {
-    rows,
-    useRegExp = false,
-    matchCase = false,
-    matchWord = false,
-  } = options;
-
-  // escape regex
-  if (!useRegExp || !isValidRegExp(searchQuery)) {
-    searchQuery = escapeRegExp(searchQuery);
-  }
-
-  return rows.filter((row): boolean => {
-    const hasExcludeFilters = Boolean(excludeFieldMatches.length);
-    const hasIncludeFilters = Boolean(includeFieldMatches.length);
-
-    if (hasExcludeFilters) {
-      const shouldExclude = excludeFieldMatches.every(([field, value]) => {
-        const cell = row.find((cell) => cell.key === field);
-        if (!cell) return false;
-
-        const passesCriteria = meetsCriteria(cell.value as string, value, {
-          matchCase,
-          matchWord: true,
-          useRegExp,
-        });
-
-        return passesCriteria;
-      });
-
-      if (shouldExclude) return false;
-    }
-
-    if (hasIncludeFilters) {
-      const shouldInclude = includeFieldMatches.every(([field, value]) => {
-        const cell = row.find((cell) => cell.key === field);
-        if (!cell) return false;
-
-        return meetsCriteria(cell.value as string, value, {
-          matchCase,
-          matchWord: true,
-          useRegExp,
-        });
-      });
-
-      if (!shouldInclude) return false;
-    }
-
-    return row.some((field): boolean => {
-      return meetsCriteria(String(field.value ?? ""), searchQuery, {
-        matchCase,
-        matchWord,
-        useRegExp,
-      });
-    });
-  });
 }
 
 export type DataType =
